\name{fundamentusR-package}
\alias{fundamentusR-package}
\alias{fundamentusR}
\docType{package}
\title{
\packageTitle{fundamentusR}
}
\description{
\packageDescription{fundamentusR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fundamentusR}
\packageIndices{fundamentusR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fundamentusR}

Maintainer: \packageMaintainer{fundamentusR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
